Move tricore from asm to anal ##arch
