+++ libr/anal/meson.build
@@ -133,6 +133,9 @@ r_anal_sources = [
   join_paths('..','asm','arch','tms320','c55x_plus','decode_funcs.c'),
   join_paths('..','asm','arch','tms320','c55x_plus','hashtable.c'),
   join_paths('..','asm','arch','tms320','c55x_plus','utils.c'),
+  join_paths('../asm/arch','tricore','gnu','cpu-tricore.c'),
+  join_paths('../asm/arch','tricore','gnu','tricore-dis.c'),
+  join_paths('../asm/arch','tricore','gnu','tricore-opc.c'),
   join_paths('..','asm','arch','v810','v810_disas.c'),
   join_paths('..','asm','arch','xtensa','gnu','xtensa-dis.c'),
   join_paths('..','asm','arch','xtensa','gnu','elf32-xtensa.c'),
