+++ libr/anal/p/tricore.mk
@@ -1,4 +1,7 @@
 OBJ_TRICORE=anal_tricore.o
+OBJ_TRICORE+=../../asm/arch/tricore/gnu/tricore-dis.o
+OBJ_TRICORE+=../../asm/arch/tricore/gnu/tricore-opc.o
+OBJ_TRICORE+=../../asm/arch/tricore/gnu/cpu-tricore.o
 
 STATIC_OBJ+=${OBJ_TRICORE}
 TARGET_TRICORE=anal_tricore.${EXT_SO}
