+++ libr/asm/meson.build
@@ -51,7 +51,6 @@ r_asm_sources = [
   join_paths('p','asm_sparc_cs.c'),
   join_paths('p','asm_sparc_gnu.c'),
   join_paths('p','asm_tms320.c'),
-  join_paths('p','asm_tricore.c'),
   join_paths('p','asm_v810.c'),
   join_paths('p','asm_v850.c'),
   join_paths('p','asm_v850_gnu.c'),
@@ -144,9 +143,6 @@ r_asm_sources = [
   join_paths('arch','tms320','c55x_plus','ins.c'),
   join_paths('arch','tms320','c55x_plus','utils.c'),
   join_paths('arch','tms320','tms320_dasm.c'),
-  join_paths('arch','tricore','gnu','cpu-tricore.c'),
-  join_paths('arch','tricore','gnu','tricore-dis.c'),
-  join_paths('arch','tricore','gnu','tricore-opc.c'),
   join_paths('arch','v810','v810_disas.c'),
   join_paths('arch','v850','v850_disas.c'),
   join_paths('arch','vax','vax-dis.c'),
